// projects/audiodoe-app/src/app/api-client/models/category/categoryTranslation.ts
export enum ECategory {
  Animals = 'Animals',
  Feelings = 'Feelings',
  Family = 'Family',
  Party = 'Party',
  History = 'History',
  Comedy = 'Comedy',
  Sports = 'Sports',
  Adventure = 'Adventure',
}

export const CategoryTranslationMap = {
  [ECategory.Animals]: 'Dieren',
  [ECategory.Feelings]: 'Gevoelens',
  [ECategory.Family]: 'Familie',
  [ECategory.Party]: 'Feesten',
  [ECategory.History]: 'Geschiedenis',
  [ECategory.Comedy]: 'Humor',
  [ECategory.Sports]: 'Sport',
  [ECategory.Adventure]: 'Avontuur',
}